oveanet-pack/seo-audit-360/extensions/google-search-console/requirements.txt
# Google Search Console Extension — Dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
